Fix URe ROS example

+++ projects/robots/universal_robots/resources/ros_package/ur_e_webots/scripts/universal_robots_ros.py
@@ -20,8 +20,8 @@ import argparse
 import rospy
 
 from controller import Robot
-from joint_state_publisher import JointStatePublisher
-from trajectory_follower import TrajectoryFollower
+from ur_e_webots.joint_state_publisher import JointStatePublisher
+from ur_e_webots.trajectory_follower import TrajectoryFollower
 from rosgraph_msgs.msg import Clock
 
 
@@ -56,5 +56,5 @@ while robot.step(timestep) != -1 and not rospy.is_shutdown():
     time = robot.getTime()
     msg.clock.secs = int(time)
     # round prevents precision issues that can cause problems with ROS timers
-    msg.clock.nsecs = round(1000 * (time - msg.clock.secs)) * 1.0e+6
+    msg.clock.nsecs = int(round(1000 * (time - msg.clock.secs)) * 1.0e+6)
     clockPublisher.publish(msg)
